requirements.txt
opencv-python
pyzbar
flask
pillow 
pymupdf
requests
flask_cors
lxml
beautifulsoup4